hotel-booking-agent-autogen-agent-iam/assistant-agent/requirements.txt
fastapi==0.115.9
uvicorn[standard]==0.34.2
autogen_agentchat==0.6.4
autogen-ext==0.6.4
python-dotenv>=0.19.0
authlib==1.5.2
openai==1.79.0
tiktoken==0.9.0
cachetools==5.5.2
requests==2.32.3
httpx>=0.28.0
pydantic>=1.8.0
PyJWT>=2.0.0
python-multipart>=0.0.5
starlette>=0.20.0
websockets>=10.0
asgardeo-ai==0.2.2
